pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git_playbook"
version = "1.1.0"
description = "git playbook"
readme = "README.md"
requires-python = ">= 3.10.9"
license = { file = "LICENSE" }
authors = [{ name = "Ankit Shrivastava" }]
keywords = [""]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
]

dependencies = [
  "numpy",
  "matplotlib",
  "jupyterlab",
  "pytest",
  "pytest-cov",
  "ruamel.yaml",
  "torchvision",
  "torchinfo",
  "tensorboard",
] # Add your pip package dependencies here

[project.scripts]
install-torch = "utils.torch_setup:install_torch_according_to_hardware"

[project.optional-dependencies]
docs = [
  "sphinx",
  "sphinx-rtd-theme",
  "nbsphinx",
  "myst-parser",
  "sphinx-last-updated-by-git",
  "sphinx-autodoc-typehints",
]
dev = [
  # --- Linting & Formatting ---
  "black==24.10.0",
  "ruff==0.7.0",
  "mypy==1.13.0",
  "pre-commit",

  # --- Debugging / REPL / Notebook ---
  "ipython==8.29.0",
  "ipykernel==6.29.5",
  "ipdb==0.13.13",
  "debugpy==1.8.7",

  # --- Profilers ---
  "line_profiler==4.1.3",
  "memory_profiler",
  "scalene==1.5.33.1",
]


# -------------------------
# mypy
# -------------------------
[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
follow_imports = "skip"
warn_unused_configs = true
warn_redundant_casts = false
warn_unused_ignores = false
disallow_untyped_defs = false
check_untyped_defs = false
no_implicit_optional = false
exclude = '(\.venv|venv|env|\.env|\.mypy_cache|build|dist|_build|data/raw|data/processed)'


# -------------------------
# (formatter)
# -------------------------
[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

# Exclude all temporary, cache, and build folders
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.ipynb_checkpoints
  | __pycache__
  | \.venv
  | venv
  | env
  | \.env
  | \.virtualenv
  | \.virtualenvs
  | build
  | dist
  | _build
  | outputs?
  | checkpoints?
  | data/raw
  | data/processed
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = [
  "E203", # Black conflict
]

exclude = [
  ".git",
  ".hg",
  ".mypy_cache",
  ".pytest_cache",
  ".ipynb_checkpoints",
  "__pycache__",
  ".venv",
  "venv",
  "env",
  ".env",
  ".virtualenv",
  ".virtualenvs",
  "_build",
  "build",
  "dist",
  "data/raw",
  "data/processed",
]

[tool.ruff.lint.isort]
known-first-party = ["mlops_playground"]

# -------------------------
# pytest configuration
# -------------------------
[tool.pytest.ini_options]
minversion = "6.0"

addopts = [
  "-v",
  "-ra",                # show summary for skipped/xfail
  "--strict-markers",   # avoid typos in markers
  "--disable-warnings", # cleaner output unless debugging
  # "--cov=src",
  # "--cov-report=term-missing",
]

testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
markers = ["slow: slow tests", "integration: integration tests"]

# -------------------------
# Coverage (pytest-cov)
# -------------------------
# [tool.coverage.run]
# branch = true
# source = ["src"]

# [tool.coverage.report]
# show_missing = true
# skip_covered = true
# precision = 1

# -------------------------
# setuptools: package discovery
# -------------------------

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pylint.'MASTER']
init-hook = 'import sys; sys.path.insert(0, "src")'
